+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 - Refresh User Updates list after Remove/Update signals.
 - Fixing segmentation fault issue after updating the packages one-by-one.
 - Updating the built-in terminal GUI.
+- Updating to better working the GUI in lower resolution.
+- Improvment in CategoryView GUI object.
 
 
 ## [0.3.2]